Refactoring of how trace messages are collected
In order to support high concurrency, kprof_tracer will now handle all trace messages in batches. At a given interval, the queue of messages is processed, searching for requests. This removes race conditions in the previous version where a request would register a handler for a process during the time of that request.